refact(ProductPage): unify layout
