+++ apps/store/src/blocks/ProductPageBlock.tsx
@@ -1,0 +1,266 @@
+import styled from '@emotion/styled'
+import { storyblokEditable, StoryblokComponent, SbBlokData } from '@storyblok/react'
+import { useState, useEffect, useRef } from 'react'
+import { theme, mq } from 'ui'
+import { GridLayout } from '@/components/GridLayout/GridLayout'
+import { useProductPageContext } from '@/components/ProductPage/ProductPageContext'
+import { PurchaseForm } from '@/components/ProductPage/PurchaseForm/PurchaseForm'
+import { ProductVariantSelector } from '@/components/ProductVariantSelector/ProductVariantSelector'
+import { SbBaseBlockProps } from '@/services/storyblok/storyblok'
+import { zIndexes } from '@/utils/zIndex'
+
+const TABLIST_HEIGHT = '2.5rem'
+
+export type PageSection = 'overview' | 'coverage'
+
+export type ProductPageBlockProps = SbBaseBlockProps<{
+  overviewLabel: string
+  coverageLabel: string
+  overview: SbBlokData[]
+  coverage: SbBlokData[]
+  body: SbBlokData[]
+}>
+
+export const ProductPageBlock = ({ blok }: ProductPageBlockProps) => {
+  const { productData } = useProductPageContext()
+  const [activeSection, setActiveSection] = useState<PageSection>('overview')
+
+  useActiveSectionChangeListener(['overview', 'coverage'], (sectionId) =>
+    setActiveSection(sectionId as PageSection),
+  )
+
+  const shouldRenderVariantSelector =
+    activeSection === 'coverage' && productData.variants.length > 1
+
+  return (
+    <Main {...storyblokEditable(blok)}>
+      <StickyHeader>
+        <>
+          <nav aria-label="page content">
+            <ContentNavigationList>
+              <li>
+                <ContentNavigationTrigger
+                  href="#overview"
+                  data-state={activeSection === 'overview' ? 'active' : 'inactive'}
+                  aria-current={activeSection === 'overview' ? 'true' : undefined}
+                >
+                  {blok.overviewLabel}
+                </ContentNavigationTrigger>
+              </li>
+              <li>
+                <ContentNavigationTrigger
+                  href="#coverage"
+                  data-state={activeSection === 'coverage' ? 'active' : 'inactive'}
+                  aria-current={activeSection === 'coverage' ? 'true' : undefined}
+                >
+                  {blok.coverageLabel}
+                </ContentNavigationTrigger>
+              </li>
+            </ContentNavigationList>
+          </nav>
+          {shouldRenderVariantSelector && <VariantSelector />}
+        </>
+      </StickyHeader>
+
+      <Grid>
+        <GridLayout.Content width="1/2" align="right">
+          <PurchaseFormWrapper>
+            <PurchaseForm />
+          </PurchaseFormWrapper>
+        </GridLayout.Content>
+
+        <OverviewGridArea width="1/2" align="left">
+          <OverviewSection id="overview">
+            {blok.overview?.map((nestedBlock) => (
+              <StoryblokComponent blok={nestedBlock} key={nestedBlock._uid} />
+            ))}
+          </OverviewSection>
+        </OverviewGridArea>
+
+        <GridLayout.Content width="1" align="center">
+          <section id="coverage">
+            {blok.coverage?.map((nestedBlock) => (
+              <StoryblokComponent blok={nestedBlock} key={nestedBlock._uid} />
+            ))}
+          </section>
+        </GridLayout.Content>
+      </Grid>
+
+      {blok.body.map((nestedBlock) => (
+        <StoryblokComponent blok={nestedBlock} key={nestedBlock._uid} />
+      ))}
+    </Main>
+  )
+}
+ProductPageBlock.blockName = 'product'
+
+const VariantSelector = () => {
+  return (
+    <VariantSelectorWrapper>
+      <StyledProductVariantSelector />
+    </VariantSelectorWrapper>
+  )
+}
+
+const StickyHeader = styled.div({
+  position: 'sticky',
+  top: theme.space.sm,
+  zIndex: zIndexes.tabs,
+  paddingInline: theme.space.md,
+
+  [mq.md]: {
+    paddingInline: theme.space.lg,
+  },
+  [mq.lg]: {
+    top: theme.space.md,
+    paddingInline: theme.space.xl,
+  },
+})
+
+const ContentNavigationList = styled.ol({
+  display: 'flex',
+  gap: theme.space.xs,
+  height: TABLIST_HEIGHT,
+})
+
+const ContentNavigationTrigger = styled.a({
+  cursor: 'pointer',
+  display: 'flex',
+  alignItems: 'flex-end',
+  justifyContent: 'center',
+  textAlign: 'center',
+  paddingInline: theme.space.md,
+  paddingBlock: theme.space.xs,
+  fontSize: theme.fontSizes.md,
+  lineHeight: theme.fontSizes.xl,
+  color: theme.colors.dark,
+  backgroundColor: theme.colors.grayTranslucent100,
+  boxShadow: '0px 1px 2px rgba(0, 0, 0, 0.15)',
+  backdropFilter: 'blur(20px)',
+  borderRadius: theme.radius.sm,
+
+  '&[data-state=active]': {
+    paddingInline: theme.space.xxxl,
+    color: theme.colors.dark,
+    backgroundColor: theme.colors.grayTranslucent300,
+  },
+
+  '@media (hover: hover)': {
+    ':hover': {
+      backgroundColor: theme.colors.gray200,
+      transition: `background-color ${theme.transitions.hover}`,
+
+      '&[data-state=active]': {
+        backgroundColor: theme.colors.grayTranslucent400,
+      },
+    },
+  },
+
+  '&:focus-visible': {
+    boxShadow: `0 0 0 2px ${theme.colors.purple500}`,
+  },
+})
+
+const VariantSelectorWrapper = styled.div({
+  width: 'fit-content',
+  minWidth: '12.5rem',
+  marginTop: theme.space.xs,
+})
+
+const StyledProductVariantSelector = styled(ProductVariantSelector)({
+  backgroundColor: theme.colors.greenFill1,
+  boxShadow: '0px 1px 2px rgba(0, 0, 0, 0.15)',
+
+  ':hover, :focus-within': {
+    backgroundColor: theme.colors.greenFill3,
+  },
+})
+
+const Main = styled.main({
+  paddingTop: theme.space.sm,
+})
+
+const Grid = styled(GridLayout.Root)({
+  // TODO: Ideally padding that are currently spcing content from the edge of the page
+  // should be added by the container.
+  paddingInline: 0,
+
+  [mq.lg]: {
+    paddingInline: 0,
+  },
+})
+
+const OverviewGridArea = styled(GridLayout.Content)({
+  // Even though Overview section comes right after PurchaseForm in the DOM order,
+  // it should be "displayed" before PurchaseForm for desktop layout.
+  [mq.lg]: {
+    gridRow: 1,
+  },
+})
+
+const OverviewSection = styled.section({
+  [mq.lg]: {
+    marginTop: `calc(-${TABLIST_HEIGHT} - ${theme.space.md})`,
+  },
+})
+
+const PurchaseFormWrapper = styled.div({
+  paddingTop: '3vw',
+
+  [mq.lg]: {
+    position: 'sticky',
+    top: 0,
+    // Scroll independently if content is too long
+    maxHeight: '100vh',
+    overflow: 'auto',
+    paddingBottom: theme.space.xl,
+    paddingTop: '6vw',
+  },
+})
+
+const useActiveSectionChangeListener = (
+  sectionsId: Array<string>,
+  callback: (activeSectionId: string) => void,
+) => {
+  // Used to determine if user is scrolling up
+  const scrollPositionRef = useRef(0)
+
+  useEffect(() => {
+    const instersectionObserverCb = (entries: IntersectionObserverEntry[]) => {
+      entries.forEach((entry) => {
+        const id = entry.target.id
+        const diff = scrollPositionRef.current - window.scrollY
+        const isScrollingUp = diff > 0
+
+        if (isScrollingUp) {
+          if (!entry.isIntersecting) {
+            const activeSectionIndex = sectionsId.findIndex((sectionId) => sectionId === id)
+            const previousSectionId = sectionsId[activeSectionIndex - 1]
+            if (previousSectionId) {
+              callback(previousSectionId)
+              scrollPositionRef.current = window.scrollY
+            }
+          }
+        } else {
+          if (entry.isIntersecting) {
+            callback(id)
+            scrollPositionRef.current = window.scrollY
+          }
+        }
+      })
+    }
+
+    const observer = new IntersectionObserver(instersectionObserverCb, {
+      // Observe only the top 15% of the screen
+      rootMargin: '0% 0% -85% 0%',
+    })
+    sectionsId.forEach((sectionId) => {
+      const element = document.getElementById(sectionId)
+      if (element) {
+        observer.observe(element)
+      }
+    })
+
+    return () => observer.disconnect()
+  }, [sectionsId, callback])
+}
