--- apps/store/src/blocks/ProductPageBlock/ProductPageBlock.types.tsx
@@ -1,1 +1,0 @@
-export type PageSection = 'overview' | 'coverage'
