+++ apps/store/src/services/storyblok/storyblok.ts
@@ -36,7 +36,7 @@ import { PerilsBlock } from '@/blocks/PerilsBlock'
 import { ProductCardBlock } from '@/blocks/ProductCardBlock'
 import { ProductDocumentsBlock } from '@/blocks/ProductDocumentsBlock'
 import { ProductGridBlock } from '@/blocks/ProductGridBlock'
-import { ProductPageBlock } from '@/blocks/ProductPageBlock/ProductPageBlock'
+import { ProductPageBlock } from '@/blocks/ProductPageBlock'
 import { ProductSlideshowBlock } from '@/blocks/ProductSlideshowBlock'
 import { ReusableBlockReference } from '@/blocks/ReusableBlockReference'
 import { SpacerBlock } from '@/blocks/SpacerBlock'
